Extract DataEngine to a separate module
To avoid possible circular dependencies.
